+++ contracts/factory/src/error.rs
@@ -5,4 +5,10 @@ use soroban_sdk::contracterror;
 #[repr(u32)]
 pub enum ContractError {
     Unauthorized = 0,
+    ConfigNotSet = 1,
+    FailedToGetAdminAddrFromStorage = 2,
+    FirstTokenMustBeSmallerThenSecond = 3,
+    LiquidityPoolVectorNotFound = 4,
+    MinStakeLessOrEqualZero = 5,
+    MinRewardTooSmall = 6,
 }