+++ contracts/factory/src/lib.rs
@@ -3,11 +3,14 @@ mod contract;
 mod error;
 mod storage;
 
-pub mod token_contract {
-    // The import will code generate:
-    // - A ContractClient type that can be used to invoke functions on the contract.
-    // - Any types in the contract that were annotated with #[contracttype].
+mod utils;
+
+#[allow(clippy::too_many_arguments)]
+pub mod lp_contract {
     soroban_sdk::contractimport!(
-        file = "../../target/wasm32-unknown-unknown/release/soroban_token_contract.wasm"
+        file = "../../target/wasm32-unknown-unknown/release/phoenix_pair.wasm"
     );
 }
+
+#[cfg(test)]
+mod tests;