+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - Stable pair: Initialize contract ([#108])
 - Stake: Added new variable TotalStaked in storage ([#94])
 - Pair: Deployment of Stake contract during initialization ([#98])
+- Factory: Contract that allows us to deploy Liquidity Pools ([#112])
 - Decimal: Implement `from_atomics` and `to_string` ([#115])
 - Curve: Implement `end` helper ([#115])
 - Phoenix: Helper library for commonly used functions, structs, etc... ([#116])
@@ -41,6 +42,7 @@ and this project adheres to
 [#98]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/98
 [#100]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/100
 [#108]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/108
+[#112]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/112
 [#115]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/115
 [#116]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/116
 [#118]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/118